# Improve delete extensions from identifiables performance