**Please check if the PR fulfills these requirements**
<!-- please use `'[x]'` to check the checkboxes, or submit the PR and then click the checkboxes -->
- [x] The commit message follows our guidelines
- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] A PR or issue has been opened in all impacted repositories (if any)


**Does this PR already have an issue describing the problem?**
<!-- If so, link to this issue using `'Fixes #XXX'` and skip the rest -->
No


**What kind of change does this PR introduce?**
<!-- Bug fix, feature, docs update, ... -->
Improve performance of delete extensions from identifiables. 


**What is the current behavior?**
<!-- You can also link to an open issue here -->
In the current behavior, we delete extensions from identifiables using a `delete (extensionName1, equipmentId1) OR (extensionName1, equipmentId2) OR ...`
This is not efficient and takes around 800ms. Postgres handle this request in ~5ms but there is an overhead somewhere that I could not identify. This is visible if one put a timer around `executeUpdate()` in `deleteExtensionsFromIdentifiables`.


**What is the new behavior (if this is a feature change)?**
We delete extensions from identifiables by grouping by extension names. One batch is done for each extensionName. The sql is like: `delete where extensionName = extensionName1 and id in (equipmentId1, equipmentId2, ....`
This is efficient and takes around 10ms. Updating 1000 generators is again around 200ms instead of 1000ms.